// StoreApp/src/types/ManifestForm.ts
// Tu type ManifestForm
export type ManifestForm = {
    fechaRecoleccion: string;
    numeroManifiesto: number;
    razonSocial: string;
    municipio: string;
    estado: string;    
    transporte: string;
    operador: string;
    fechaEntrega: string;
    numeroPlacas: number;
    recat: number;
    fechaEntregaSucursal: string;
    observaciones: string;
};